requirements.txt
yt-dlp
numpy
torch
whisperx
pydub
resemblyzer
scikit-learn
pyyaml
openai
python-dotenv
